requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==4.0.2
contextlib2==0.6.0.post1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
enum34==1.1.9
filelock==3.0.12
Flask==1.1.2
idna==2.10
importlib-metadata==1.5.0
importlib-resources==1.4.0
iniconfig==1.0.1
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.3
jupyterlab==0.33.12
jupyterlab-launcher==0.11.2
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==4.4.0
notebook==5.7.8
packaging==20.4
pandocfilters==1.4.2
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==1.0.18
ptyprocess==0.6.0
py==1.9.0
Pygments==2.5.2
PyHamcrest==2.0.2
pyparsing==2.4.7
pyrsistent==0.15.7
pyserial==3.4
pytest==6.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pyzmq==19.0.0
qtconsole==4.6.0
requests==2.24.0
scandir==1.10.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
subprocess32==3.5.4
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==5.1.1
traitlets==4.3.3
typing==3.7.4.1
urllib3==1.25.10
virtualenv==20.0.13
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==1.2.0
